:sparkles: add ban repository: now just rawquery #152
add ban repository: now just rawquery #152
